+++ packages/cli/__tests__/zosfiles/__integration__/compare/ds/__scripts__/command/command_compare_data_set.sh
@@ -1,0 +1,12 @@
+#!/bin/bash
+dsn=$1
+dsn2=$2
+opt1=$3
+set -e
+
+echo "================Z/OS FILES COMPARE DATA SET==============="
+zowe files compare ds "$dsn" "$dsn2" $opt1
+if [ $? -gt 0 ]
+then
+    exit $?
+fi