+++ packages/cli/__tests__/zosfiles/__integration__/compare/ds/__scripts__/command_compare_data_set_help.sh
@@ -1,0 +1,13 @@
+#!/bin/bash
+set -e
+
+echo "================Z/OS FILES LIST DATA SET HELP==============="
+zowe zos-files compare ds --help
+if [ $? -gt 0 ]
+then
+    exit $?
+fi
+
+echo "================Z/OS FILES LIST DATA SET HELP WITH RFJ==========="
+zowe zos-files compare ds --help --rfj
+exit $?