+++ packages/cli/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to the Zowe CLI package will be documented in this file.
 
+## `7.4.0`
+
+- Enhancement: Added the `zowe zos-files compare data-set` command to compare two datasets and display the differences on the terminal. [#1442](https://github.com/zowe/zowe-cli/issues/1442)
+- BugFix: Alter the `zowe daemon disable` command to only kill the daemon running for the current user.
+
 ## `7.3.1`
 
 - BugFix: Updated Imperative to fix CLI commands failing with error "Cannot find module 'ansi-colors'".